requirements.txt
blinker==1.4
click==7.1.1
feedparser==5.2.1
Flask==1.1.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==1.4.6
Werkzeug==1.0.0
